♻️  zb: Split server handshake code into its own mod
